automation/e2e/tucana_automation/pyproject.toml
[tool.poetry]
name = "build_qa_automation"
version = "0.1.0"
description = "Utilities for building the automation"
authors = ["Guillermo Oliva <guilleoliva@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
httpx = "^0.25.2"
hl7 = "^0.4.5"
aiorun = "^2023.7.2"
typer = "^0.12.3"

[tool.poetry.dev-dependencies]
webdriver-manager = "==4.0.1"
onetimepass = "==1.0.1"
selenium = "==3.141.0"
pandas = "*"
behave = "==1.2.6"
beautifulsoup4 = "==4.12.2"
behavex="==1.6.0"
behaving="==3.0.4"
jsondiff = "*"
genson = "*"
pyshadow = "*"
configparser = "~=5.2.0"
numpy = "==1.24.2"
python-dateutil = "~=2.8.2"
requests = "~=2.31.0"
jsonschema = "*"
black = "^22.8.0"
reformat-gherkin="*"
isort="*"
pydantic="*"
websocket-client="==1.5.1"
behavex-web = {path = "./bin/behavex-web-1.3.6.zip"}
pyotp = "*"
urllib3 = "~=1.26"
faker = "*"
Pillow = "==10.2.0"
certifi = "~=2023.7.22"

[tool.poetry.group.dev.dependencies]
polyfactory = "^2.12.0"

